+++ src/main/java/com/liutao/demo/domain/BaseBean.java
@@ -1,0 +1,12 @@
+package com.liutao.demo.domain;
+
+
+import com.alibaba.fastjson.JSONObject;
+
+public class BaseBean {
+
+    @Override
+    public String toString() {
+        return JSONObject.toJSONString(this);
+    }
+}