+++ src/main/java/com/liutao/demo/domain/User.java
@@ -1,17 +1,12 @@
 package com.liutao.demo.domain;
 
-import java.util.Date;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
 import com.fasterxml.jackson.annotation.JsonFormat;
 
+import javax.persistence.*;
+import java.util.Date;
+
 @Entity(name="user")
-public class User {
+public class User extends BaseBean{
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)