+++ src/main/resources/static/js/lib/videojs/CHANGELOG.md
@@ -100,7 +100,7 @@ _(none)_
 * Fixed a bug with setPoster() in the minified version ([view](https://github.com/videojs/video.js/pull/1062))
 
 ## 4.4.2 (2014-02-24)
-* Fixed module.exports in minified version ([view](https://github.com/videojs/video.js/pull/1038))
+* Fixed com.liutao.demo.com.elead.ppm.module.exports in minified version ([view](https://github.com/videojs/video.js/pull/1038))
 
 ## 4.4.1 (2014-02-18)
 * Added .npmignore so dist files wouldn't be ignored in packages